refactor: jwt -> security directory
