+++ backend/core/src/main/java/com/rollthedice/backend/global/security/jwt/exception/NotFoundTokenException.java
@@ -1,4 +1,4 @@
-package com.rollthedice.backend.global.jwt.exception;
+package com.rollthedice.backend.global.security.jwt.exception;
 
 public class NotFoundTokenException extends RuntimeException {
     public NotFoundTokenException() {
